Agents/wrds_agent/requirements.txt
# Core dependencies
pydantic>=1.9.0
sqlalchemy>=1.4.0
pandas>=2.0.0
psycopg2-binary>=2.9.6
python-dotenv>=1.0.0
wrds>=3.1.0

# OpenAI and AI tools
openai>=1.3.0

# Logging and monitoring
loguru>=0.7.0

# Data visualization (optional)
matplotlib>=3.5.0
tqdm>=4.66.1
